refactor: Separate Entity List component
